architecture checking better handled, continue to write struct printer

+++ inc/type.h
@@ -7,54 +7,54 @@
 
 struct user_regs_struct64
 {
-	__extension__ unsigned long long int r15;
-	__extension__ unsigned long long int r14;
-	__extension__ unsigned long long int r13;
-	__extension__ unsigned long long int r12;
-	__extension__ unsigned long long int rbp;
-	__extension__ unsigned long long int rbx;
-	__extension__ unsigned long long int r11;
-	__extension__ unsigned long long int r10;
-	__extension__ unsigned long long int r9;
-	__extension__ unsigned long long int r8;
-	__extension__ unsigned long long int rax;
-	__extension__ unsigned long long int rcx;
-	__extension__ unsigned long long int rdx;
-	__extension__ unsigned long long int rsi;
-	__extension__ unsigned long long int rdi;
-	__extension__ unsigned long long int orig_rax;
-	__extension__ unsigned long long int rip;
-	__extension__ unsigned long long int cs;
-	__extension__ unsigned long long int eflags;
-	__extension__ unsigned long long int rsp;
-	__extension__ unsigned long long int ss;
-	__extension__ unsigned long long int fs_base;
-	__extension__ unsigned long long int gs_base;
-	__extension__ unsigned long long int ds;
-	__extension__ unsigned long long int es;
-	__extension__ unsigned long long int fs;
-	__extension__ unsigned long long int gs;
+	__extension__ unsigned long long int	r15;
+	__extension__ unsigned long long int	r14;
+	__extension__ unsigned long long int	r13;
+	__extension__ unsigned long long int	r12;
+	__extension__ unsigned long long int	rbp;
+	__extension__ unsigned long long int	rbx;
+	__extension__ unsigned long long int	r11;
+	__extension__ unsigned long long int	r10;
+	__extension__ unsigned long long int	r9;
+	__extension__ unsigned long long int	r8;
+	__extension__ unsigned long long int	rax;
+	__extension__ unsigned long long int	rcx;
+	__extension__ unsigned long long int	rdx;
+	__extension__ unsigned long long int	rsi;
+	__extension__ unsigned long long int	rdi;
+	__extension__ unsigned long long int	orig_rax;
+	__extension__ unsigned long long int	rip;
+	__extension__ unsigned long long int	cs;
+	__extension__ unsigned long long int	eflags;
+	__extension__ unsigned long long int	rsp;
+	__extension__ unsigned long long int	ss;
+	__extension__ unsigned long long int	fs_base;
+	__extension__ unsigned long long int	gs_base;
+	__extension__ unsigned long long int	ds;
+	__extension__ unsigned long long int	es;
+	__extension__ unsigned long long int	fs;
+	__extension__ unsigned long long int	gs;
 };
 
 struct user_regs_struct32
 {
-	long int ebx;
-	long int ecx;
-	long int edx;
-	long int esi;
-	long int edi;
-	long int ebp;
-	long int eax;
-	long int xds;
-	long int xes;
-	long int xfs;
-	long int xgs;
-	long int orig_eax;
-	long int eip;
-	long int xcs;
-	long int eflags;
-	long int esp;
-	long int xss;
+	uint32_t	ebx;
+	uint32_t	ecx;
+	uint32_t	edx;
+	uint32_t	esi;
+	uint32_t	edi;
+	uint32_t	ebp;
+	uint32_t	eax;
+	uint32_t	xds;
+	uint32_t	xes;
+	uint32_t	xfs;
+	uint32_t	xgs;
+	uint32_t	orig_eax;
+	uint32_t	eip;
+	uint32_t	xcs;
+	uint32_t	eflags;
+	uint32_t	esp;
+	uint32_t	xss;
 };
